Fix DNS resolution
Fixed resolution of names when a custom DNS server address is provided in the
NSE response. When a custom dns endpoint with a specific search domain is provided, there
would be two DNS servers:
kube-dns to resolve cluster.local names and the custom dns to resolve custom
domain names.

The NSM dnsutils code sends DNS queries to both the servers simultaneously
and returns prematurely if one of the servers sends a response with
zero answers, leading to an empty response to the querier.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>