fix(): Fix dns caching
The cache chain in the dns proxy server caches responses with no data in
the answer section but the ttl update and validation functions expect
the answer section to be present in the cached entries. This causes
infinite caching of entries without any answer section and dns queries
returned with empty responses without the queries being forwarded to the
upstream dns servers.

Added a conditional to check for the availability of the Answer section
before storing the response in the cache.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>