fix(): Support tcp connection to nsmgr (#7)
* fix(): Support tcp connection to nsmgr

Added code to connect to nsmgr using a tcp socket. An important
part of the implementation was to resolve the dns name of the nsmgr
service. The resolution needs to happen before the cmd-nsc connects to
the nsmgr. Since the resolv.conf in the cmd-nsc container is overwritten to
point to the local host and the dns proxy is not configured with the IP
addresses of the upstream dns servers, the resolution uses the IP
address of the kube-dns server stored in the original resolv.conf at:
/etc/nsm-dns-config/resolv.conf.restore.

With this change, the cmd-nsc can connect to nsmgr using either a unix
domain socket or a tcp socket.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>

* fix(): Support tcp connection to nsmgr

Fixed code formatting.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>

---------

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>