fix(): Check pod network before connection to nsmgr
Added a check if pod network is ready before making connection to the nsmgr over tcp.
This is needed if the cmd-nsc sidecar is running alongside the istio-proxy sidecar.
If istio is enabled on the pod, the istio-init container installs iptable rules to redirect
all incoming and outgoing traffic to the port numbers that the istio-proxy listens on.
This leads to a condition where the pod network is virtually dead from the time istio-init
installs the iptable rules to the time the istio-proxy sidecar boots up and is ready to listen
on the port numbers to which all the traffic is redirected. This means that any other container
in the pod cannot make network connections to the outside world until the istio-proxy is ready.
This causes the cmd-nsc to crashloop trying to reach nsmgr over tcp. So we need to check if the
pod network is operational before attempting to connect to the nsmgr.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>