fix(): Fixed dns latency issues
DNS resolution through the proxy in the cmd-nsc sidecar was
taking too long because the proxy waited for responses for all
FQDNs generated using the search domain list in the resolv.conf before
returning a response to the caller.
Added a fix to return the fix successful response without waiting for
all other FQDNs.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>